ma fonction de sleep ne marche pas

+++ rank03/philosophers/philo/src/utils/init.c
@@ -6,7 +6,7 @@
 /*   By: bverdeci <bverdeci@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/20 17:11:00 by bverdeci          #+#    #+#             */
-/*   Updated: 2023/06/20 17:57:53 by bverdeci         ###   ########.fr       */
+/*   Updated: 2023/06/21 14:23:03 by bverdeci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,9 +47,9 @@ int	init_table(t_table *table, t_philo **philos, char **av)
 	if (av[5])
 		table->foods = my_atoi(av[5]);
 	table->philos = *philos;
-	table->t_die = my_atoi(av[2]);
-	table->t_eat = my_atoi(av[3]);
-	table->t_sleep = my_atoi(av[4]);
+	table->t_die = (long int)my_atoi(av[2]);
+	table->t_eat = (long int)my_atoi(av[3]);
+	table->t_sleep = (long int)my_atoi(av[4]);
 	return (0);
 }
